ZZZ_embedding/requirements.txt
# Core dependencies
openai>=1.0.0
azure-search-documents>=11.4.0
tiktoken>=0.5.0
numpy>=1.24.0
pandas>=2.0.0

# Azure services
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-core>=1.29.0

# Graph database
gremlin_python>=3.7.0

# Async support
aiohttp>=3.8.0
asyncio>=3.4.3

# Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0 